pyproject.toml
[project]
name = "synth-engine"
version = "0.2.0"
description = "Clinically anchored multi-modal synthesis engine with symbolic overlays and family constellation analytics."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7",
  "python-jose[cryptography]>=3.3",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "numpy>=1.24",
  "networkx>=3.2",
  "pyswisseph>=2.10.3.2",
]
